+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 -----------
 
 ### Internals
-* None.
+* Realm::write_copy_without_client_file_id and Realm::write_copy are merged. For synchronized realms, the file written will have the client file ident removed. Furthermore it is required that all local changes are synchronized with the server before the copy can. The function will throw if there are pending uploads.
 
 ----------------------------------------------